# docs/swagger.yaml
info:
  contact: {}
paths:
  /contact/addfriend:
    post:
      parameters:
      - description: userId
        in: formData
        name: userId
        type: string
      - description: targetId
        in: formData
        name: userId
        type: string
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 创建群聊
      tags:
      - 聊天模块
  /contact/createCommunity:
    post:
      parameters:
      - description: ownerId
        in: formData
        name: ownerId
        type: string
      - description: name
        in: formData
        name: Name
        type: string
      - description: icon
        in: formData
        name: Img
        type: string
      - description: desc
        in: formData
        name: Desc
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 创建群聊
      tags:
      - 聊天模块
  /contact/joinGroup:
    post:
      parameters:
      - description: userId
        in: formData
        name: userId
        type: string
      - description: comId
        in: formData
        name: comId
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 加入群聊
      tags:
      - 聊天模块
  /contact/loadcommunity:
    post:
      parameters:
      - description: ownerId
        in: formData
        name: ownerId
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 加载群列表
      tags:
      - 聊天模块
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /searchFriends:
    post:
      parameters:
      - description: userId
        in: formData
        name: userId
        type: string
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 加载好友
      tags:
      - 聊天模块
  /user/createUser:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户模块
  /user/deleteUser:
    post:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/findUserByNameAndPwd:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 名字密码查询用户
      tags:
      - 用户模块
  /user/getUserList:
    get:
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 获取用户列表
      tags:
      - 用户模块
  /user/redisMsg:
    post:
      parameters:
      - description: userIdA
        in: formData
        name: userIdA
        type: string
      - description: userIdB
        in: formData
        name: userIdB
        type: string
      - description: start
        in: formData
        name: start
        type: string
      - description: end
        in: formData
        name: end
        type: string
      - description: isRev
        in: formData
        name: isRev
        type: boolean
      responses:
        "200":
          description: code", "message", "data}
          schema:
            type: string
      summary: 载入聊天缓存
      tags:
      - 聊天模块
  /user/sendMsg:
    get:
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 发送消息
      tags:
      - 聊天模块
  /user/sendUserMsg:
    get:
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 用户发送
      tags:
      - 聊天模块
  /user/updateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code", "message", "data"}
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
swagger: "2.0"